pyproject.toml
[project]
name = "zube-mcp"
version = "0.2.0"
description = "MCP server for the Zube.io project management API"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "httpx>=0.27",
    "PyJWT>=2.8",
    "cryptography>=42.0",
]

[project.scripts]
zube-mcp = "zube_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
